change to double

+++ src/main/java/Calculator.java
@@ -1,5 +1,5 @@
 public class Calculator {
-    public int multiply(int a, int b) {
-        return a * b;
+    public double multiply(double a, double b) {
+        return a*b;
     }
 }
